tests/wpt/web-platform-tests/speculation-rules/prefetch/redirect-url.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/dispatcher/dispatcher.js"></script>
<script src="/common/utils.js"></script>
<script src="resources/utils.sub.js"></script>

<script>
  promise_test(async t => {
    assert_implements(HTMLScriptElement.supports('speculationrules'), "Speculation Rules not supported");

    let url = getRedirectUrl();
    insertSpeculationRules({ prefetch: [{ source: 'list', urls: [url] }] });
    await new Promise(resolve => t.step_timeout(resolve, 2000));

    assert_equals(await isUrlPrefetched(url), 1, "redirected url should be prefetched");
  }, "browser should be able to prefetch redirected urls");
</script>
